:art: refactor: 기존 컴포넌트 경로 변경
